Test for dtype

+++ ivy_tests/test_ivy/test_frontends/test_jax/test_jax_numpy_namespace_functions.py
@@ -700,7 +700,7 @@ def test_jax_numpy_cos(
 @handle_cmd_line_args
 @given(
     dtype_and_x=helpers.dtype_and_values(available_dtypes=helpers.get_dtypes("float")),
-    dtype=helpers.get_dtypes("float", full=False, none=True),
+    dtype=helpers.get_dtypes("numeric", full=False, none=True),
     num_positional_args=helpers.num_positional_args(
         fn_name="ivy.functional.frontends.jax.numpy.cosh"
     ),
